wip: new cfg validator
